wallet connect hook, to fix multiple state updates autologin
