+++ src/pages/_app.tsx
@@ -14,7 +14,6 @@ import Sidebar from '@/components/common/Sidebar'
 import Viewport from '@/components/common/Viewport'
 import Content from '@/components/common/Content'
 import Loading from './loading'
-import { WalletConnectProvider } from '@/contexts/walletConnect'
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
@@ -22,7 +21,6 @@ export default function App({ Component, pageProps }: AppProps) {
       <GlobalStyles />
       <GlobalStylesOverride />
       <AppStateProvider>
-        <WalletConnectProvider>
           <NotificationProvider max={10} timeout={2000}>
             <Viewport>
               <Sidebar />
@@ -36,7 +34,6 @@ export default function App({ Component, pageProps }: AppProps) {
               </Main>
             </Viewport>
           </NotificationProvider>
-        </WalletConnectProvider>
       </AppStateProvider>
     </ThemeProvider>
   )
